# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food Palace Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /products/{id}/variants:
    get:
      operationId: getProductVariants
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
    post:
      operationId: createProductVariant
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /products/{id}/addons:
    get:
      operationId: getProductAddons
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductAddon"
    post:
      operationId: createProductAddon
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductAddonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductAddon"

  # ---- DELIVERY ZONES ----
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (user gets own, admin gets all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/payment:
    post:
      operationId: confirmPayment
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ---- FAVORITES ----
...
components:
  schemas:
...
    ProductVariant:
      type: object
      required: [id, productId, name, price, isAvailable]
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    ProductVariantInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    ProductAddon:
      type: object
      required: [id, productId, name, price, isAvailable]
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    ProductAddonInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
    Product:
      type: object
      required: [id, categoryId, name, slug, basePrice, isAvailable, isFeatured, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        basePrice:
          type: number
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/ProductAddon"
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [categoryId, name, slug, basePrice]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        basePrice:
          type: number
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
    ProductUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        basePrice:
          type: number
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer

    # ---- ZONE SCHEMAS ----
...
    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        variantName:
          type: ["string", "null"]
        addonNames:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    OrderItemInput:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        variantId:
          type: integer
        variantName:
          type: string
        addonIds:
          type: array
          items:
            type: integer
        addonNames:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: string
    Order:
      type: object
      required: [id, status, paymentMethod, paymentStatus, subtotal, deliveryFee, total, deliveryAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        deliveryLocation:
          type: ["string", "null"]
        zoneId:
          type: ["integer", "null"]
        deliveryNotes:
          type: ["string", "null"]
        estimatedDeliveryTime:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [customerName, customerPhone, paymentMethod, subtotal, deliveryFee, total, deliveryAddress, items]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        paymentMethod:
          type: string
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        deliveryLocation:
          type: string
        zoneId:
          type: integer
        deliveryNotes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        estimatedDeliveryTime:
          type: string
    PaymentConfirmInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
        paymentStatus:
          type: string

    # ---- FAVORITES ----
...