# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  customerName?: string;
  customerPhone?: string;
  /** @nullable */
  customerEmail?: string | null;
  status: string;
  paymentMethod: string;
  paymentStatus: string;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress: string;
  /** @nullable */
  deliveryLocation?: string | null;
  /** @nullable */
  zoneId?: number | null;
  /** @nullable */
  deliveryNotes?: string | null;
  /** @nullable */
  estimatedDeliveryTime?: string | null;
  items?: OrderItem[];
  createdAt: Date;
  updatedAt?: Date;
}