# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  categoryId: number;
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  basePrice: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
}